test_app/requirements.txt
django-templated-email>=0.4.7,<0.5
factory_boy>=1.3
mock>=1.0.0
django-discover-runner>=0.3